Update PSMicrosoftTeams.psm1
Added parameters for detailTitle and messageSummary. Created a parameter set for messageSummary and messageBody so those two parameters are mutually exclusive. Adding that gives the ability to effectively 'hide' the messageBody in the card that displays in the Teams channel. 

Also added a helper function called Escape-JSONString to replace powershell escape characters with those that would work for JSON and preserve formatting. IF markdown or html is used in the future, this may no longer be necessary to preserve the formatting.